refactor: restructure and rename
